refactored geoqgis.py file
